app/requirements.txt
anyio==4.0.0
arabic-reshaper==3.0.0
asgiref==3.7.2
asn1crypto==1.5.1
async-timeout==4.0.3
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
Brotli==1.1.0
certifi==2023.7.22
cffi==1.16.0
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==3.3.2
click==8.1.7
constantly==23.10.4
crispy-bootstrap4==2023.1
cryptography==41.0.5
cssselect2==0.7.0
daphne==4.0.0
Django==4.2.7
django-admin-list-filter-dropdown==1.0.3
django-admin-multi-select-filter==1.3.0
django-admin-rangefilter==0.11.2
django-channels==0.7.0
django-cors-headers==4.3.0
django-crispy-forms==2.1
django-csp==3.7
django-daterange-filterspec==2.0.5
django-daterangefilter==1.0.0
django-filter==23.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-nested-routers==0.93.4
drf-yasg==1.21.7
environs==9.5.0
exceptiongroup==1.1.3
fonttools==4.46.0
gunicorn==21.2.0
h11==0.14.0
html5lib==1.1
httpcore==1.0.2
httptools==0.6.1
httpx==0.25.2
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
inflection==0.5.1
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
lxml==4.9.3
marshmallow==3.20.1
msgpack==1.0.7
oauthlib==3.2.2
oscrypto==1.3.0
packaging==23.2
Pillow==10.1.0
psycopg2==2.9.9
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydyf==0.8.0
pyHanko==0.20.1
pyhanko-certvalidator==0.24.1
PyJWT==2.8.0
pyOpenSSL==23.3.0
PyPDF3==1.0.6
pyphen==0.14.0
pypng==0.20220715.0
python-bidi==0.4.2
python-dotenv==1.0.0
python-magic==0.4.27
python-telegram-bot==20.7
pytz==2023.3.post1
PyYAML==6.0.1
qrcode==7.4.2
redis==5.0.1
referencing==0.30.2
reportlab==4.0.7
requests==2.31.0
requests-oauthlib==1.3.1
rpds-py==0.12.0
service-identity==23.1.0
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
svglib==1.5.1
tinycss2==1.2.1
tqdm==4.66.1
Twisted==23.10.0
txaio==23.1.1
typing_extensions==4.8.0
tzlocal==5.2
uritemplate==4.1.1
uritools==4.0.2
urllib3==2.0.7
uvicorn==0.24.0.post1
uvloop==0.19.0
watchfiles==0.21.0
weasyprint==60.1
webencodings==0.5.1
websockets==12.0
xhtml2pdf==0.2.8
zope.interface==6.1
zopfli==0.2.3
django-debug-toolbar
django-redis==5.4.0